requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose==3.3.0 
passlib==1.7.4       
cryptography==42.0.2  
pydantic==2.6.0
pydantic-settings==2.1.0
pytest==7.4.0
pytest-asyncio==0.23.5
httpx==0.26.0
python-multipart==0.0.6  
python-dotenv==1.0.0
asyncpg==0.30.0
tronpy==0.6.1
pyyaml==6.0.2